+++ lib/td_bg_web/views/comment_view.ex
@@ -11,11 +11,13 @@ defmodule TdBgWeb.CommentView do
   end
 
   def render("comment.json", %{comment: comment}) do
-    %{id: comment.id,
+    %{
+      id: comment.id,
       resource_id: comment.resource_id,
       resource_type: comment.resource_type,
-      created_at: comment.created_at,
+      created_at: comment.inserted_at,
       user: comment.user,
-      content: comment.content}
+      content: comment.content
+    }
   end
 end