+++ priv/repo/migrations/20200608171358_alter_comments_timestamps.exs
@@ -1,0 +1,10 @@
+defmodule TdBg.Repo.Migrations.AlterCommentsTimestamps do
+  use Ecto.Migration
+
+  def change do
+    alter table(:comments) do
+      remove(:created_at, :utc_datetime_usec)
+      remove(:updated_at, :utc_datetime_usec)
+    end
+  end
+end